+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # RELEASE HISTORY
 
 ### 0.0.7, / May, 28 2019
+* fixed bug in the CacheService methods when key contains spaces
+* added the segmentQuery property to the DeliveryOptions
+* removed the device registration methods that don't take deviceToken as parameter
+* removed the `refreshDeviceToken` method
 * customized Serialization/Deserialization for next classes: ObjectProperty, BackendlessFileInfo, GeoCategory, BackendlessUser, UserProperty
 ```
 init(from decoder: Decoder) throws
+
 func encode(to encoder: Encoder) throws
 ```
-* fixed bug in the CacheService methods when key contains spaces
-* added the segmentQuery property to the DeliveryOptions
 
 ### 0.0.6, / May, 21 2019
 * added functions to FileService: